big refactor. move extension internals into unnamed namespace. use mainwindow instead of dialog for extension window. and other renames/bugfixes
