+++ GUI/qtcommon.h
@@ -3,4 +3,6 @@
 #include "common.h"
 #include <QString>
 #include <QVector>
+#include <QMainWindow>
+#include <QFile>
 #include <QRegularExpression>
